+++ src/helm/benchmark/run_specs/enterprise_run_specs.py
@@ -76,9 +76,9 @@ def get_financial_phrasebank_spec(agreement: int = 50) -> RunSpec:
 
 
 @run_spec_function("conv_fin_qa_calc")
-def get_conv_fin_qa_spec() -> RunSpec:
+def get_conv_fin_qa_calc_spec() -> RunSpec:
     scenario_spec = ScenarioSpec(
-        class_name="helm.benchmark.scenarios.conv_fin_qa_calc_scenario.ConvFinCalcQAScenario", args={}
+        class_name="helm.benchmark.scenarios.conv_fin_qa_calc_scenario.ConvFinQACalcScenario", args={}
     )
 
     adapter_spec = get_generation_adapter_spec(input_noun="Passage", output_noun="Answer")
