+++ src/helm/benchmark/scenarios/conv_fin_qa_calc_scenario.py
@@ -1,10 +1,10 @@
 import json
 import os
-from typing import Dict, List, Tuple, Any, Optional
 import re
+from typing import Dict, List, Tuple, Any, Optional
 
-from helm.common.general import ensure_file_downloaded, ensure_directory_exists
-from .scenario import (
+from helm.benchmark.scenarios.scenario import (
+    Input,
     Scenario,
     Instance,
     Reference,
@@ -14,6 +14,7 @@ from .scenario import (
     PassageQuestionInput,
     Output,
 )
+from helm.common.general import ensure_file_downloaded, ensure_directory_exists
 
 
 def _strip_string(str: str) -> Any:
@@ -51,56 +52,31 @@ def float_equiv(str1: Optional[str], str2: Optional[str], eps: float = 1e-6) -> 
 
 
 class ConvFinQACalcScenario(Scenario):
-    """A mathematical reasoning benchmark based on the ConvFinQA paper.
+    """"A mathematical calculation benchmark based on ConvFinQA.
 
+    Data source:
+    https://github.com/czyssrs/ConvFinQA
 
-        Prompt:
-    Passage: Table:
-    {Table}
-    Text:
-    Questions:  Question: {Question}? The answer is {Answer}
-    {Question}? The answer is {Answer}
-    {Question}? The answer is {Answer}
-    {Question}? The answer is
-    Answer:
+    Reference:
+    Zhiyu Chen, Shiyang Li, Charese Smiley, Zhiqiang Ma, Sameena Shah, and William Yang Wang. 2022.
+    ConvFinQA: Exploring the Chain of Numerical Reasoning in Conversational Finance Question Answering.
+    In Proceedings of the 2022 Conference on Empirical Methods in Natural Language Processing,
+    pages 6279–6292, Abu Dhabi, United Arab Emirates. Association for Computational Linguistics.
+    https://aclanthology.org/2022.emnlp-main.421
+    """  # noqa: E501
 
-        Data source:
-        https://github.com/czyssrs/ConvFinQA
-
-        Reference:
-        Zhiyu Chen, Shiyang Li, Charese Smiley, Zhiqiang Ma, Sameena Shah, and William Yang Wang. 2022.
-        ConvFinQA: Exploring the Chain of Numerical Reasoning in Conversational Finance Question Answering.
-        In Proceedings of the 2022 Conference on Empirical Methods in Natural Language Processing,
-        pages 6279–6292, Abu Dhabi, United Arab Emirates. Association for Computational Linguistics.
-        https://aclanthology.org/2022.emnlp-main.421
-
-    """  # noqa
-
-    """ Information on this class"""
     name = "conv_fin_qa_calc"
-    description = "A mathematical calculation benchmark based on the ConvFinQA paper"
+    description = "A mathematical calculation benchmark based on ConvFinQA: Exploring the Chain of Numerical Reasoning in Conversational Finance Question Answering [(Chen ey al., 2022)](https://arxiv.org/pdf/2210.03849.pdf)."  # noqa: E501
     tags = ["question_answering", "finance"]
 
     """ Class variables """
     # Dataset file name
     DATASET_DOWNLOAD_URL: str = "https://github.com/czyssrs/ConvFinQA/raw/main/data.zip"
-    DATASET_FILE_NAME = "ConvFinQA"
 
-    def __init__(self):
-        super().__init__()
-
-    def download_dataset(self, output_path: str):
-        """Downloads the con_fin_qa dataset."""
-
-        # Download the raw data
-        data_dir = os.path.join(output_path, "data")
-        ensure_directory_exists(data_dir)
-        ensure_file_downloaded(
-            source_url=self.DATASET_DOWNLOAD_URL,
-            target_path=os.path.join(data_dir, self.DATASET_FILE_NAME),
-            unpack=True,
-            unpack_type="unzip",
-        )
+    _SPLIT_TO_JSON_FILE_NAME: Dict[str, str] = {
+        TRAIN_SPLIT: "train_turn.json",
+        VALID_SPLIT: "dev_turn.json"
+    }
 
     def get_table_text(self, table: List[List[str]]) -> str:
         """table in the format of List of columns"""
@@ -123,7 +99,7 @@ class ConvFinQACalcScenario(Scenario):
 
         return markdown
 
-    def get_instance_dict(self, dic, sep: str = "\n") -> Dict[str, Any]:
+    def get_instance_dict(self, dic, split: str, sep: str = "\n") -> Instance:
         linearized_table = self.make_pseudo_markdown_table(dic["table"], line_sep=sep)
 
         if "gold_ind" in dic["annotation"]:
@@ -145,14 +121,15 @@ class ConvFinQACalcScenario(Scenario):
                 context += q + " The answer is "
         doc = f"Table: {sep}{linearized_table}{sep}Text: {text}{sep}Questions: "
         answer = str(dic["annotation"]["exe_ans"])
-        return {
-            "input": PassageQuestionInput(passage="".join(doc), question=context, separator=" "),
-            "references": [Reference(Output(text=answer), tags=[CORRECT_TAG])],
-        }
+        return Instance(
+            input=Input(text="".join(doc) + " " + context),
+            references=[Reference(Output(text=answer), tags=[CORRECT_TAG])],
+            split=split,
+        )
 
-    def load_dataset(self, output_path: str) -> Tuple[List[Dict[str, Any]], List[Dict[str, Any]]]:
+    def load_dataset(self, output_path: str) -> Tuple[List[Instance], List[Instance]]:
         """Loads the dataset downloaded in download_dataset()."""
-        folder_path = os.path.join(output_path, "data", self.DATASET_FILE_NAME)
+        folder_path = os.path.join(output_path, "data")
         train_data = []
         dev_data = []
 
@@ -160,26 +137,25 @@ class ConvFinQACalcScenario(Scenario):
             train_raw_data = json.load(f)
 
         for problem in train_raw_data:
-            train_data.append(self.get_instance_dict(problem))
+            train_data.append(self.get_instance_dict(problem, TRAIN_SPLIT))
 
         with open(os.path.join(folder_path, "dev_turn.json"), encoding="utf-8") as f:
             dev_raw_data = json.load(f)
 
         for problem in dev_raw_data:
-            dev_data.append(self.get_instance_dict(problem))
+            dev_data.append(self.get_instance_dict(problem, VALID_SPLIT))
 
         return train_data, dev_data
 
     def get_instances(self, output_path: str) -> List[Instance]:
-        """Returns the instances for this scenario."""
-        # Body of the function
-        self.download_dataset(output_path)
+        ensure_file_downloaded(
+            source_url=self.DATASET_DOWNLOAD_URL,
+            target_path=os.path.join(output_path, "data"),
+            unpack=True,
+            unpack_type="unzip",
+        )
+
         train_data, dev_data = self.load_dataset(output_path)
-        train_k = 5
-        train_instances = [
-            Instance(input=d["input"], references=d["references"], split=TRAIN_SPLIT) for d in train_data[:train_k]
-        ]
-        valid_instances = [
-            Instance(input=d["input"], references=d["references"], split=VALID_SPLIT) for d in dev_data[:1000]
-        ]
-        return train_instances + valid_instances
+        print(len(train_data))
+        print(len(dev_data))
+        return train_data + dev_data
